atajo venta directa, decimales, mostrar tiene IVA producto

+++ src/public/js-app/compras/compras.init.js
@@ -34,7 +34,7 @@ function loadTablacompras(){
                             }
         },
         { "data": "precio" , render: function(data, type, row){
-                                return  parseInt(data).toFixed(2);
+                                return  parseFloat(data).toFixed(2);
                             }
         },
         { "data": "fecha_compra" },
